+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## main
 
+### Bug fix
+
+- Variant constructor names for union members are now allowed to be anything, since they are never exposed out to GraphQL anyway (only the typename of the object type payload is).
+
 ## 0.3.0
 
 ### New Features